+++ src/sll_ext/debug.c
@@ -3,12 +3,13 @@
 
 
 
-static sll_object_type_t _debug_type=0;
+static sll_object_type_t _debug_cs_type=0;
+static sll_object_type_t _debug_vm_cfg_type=0;
 
 
 
 static sll_object_t* _debug_get_call_stack(const sll_object_t*const* al,sll_arg_count_t all){
-	if (!_debug_type){
+	if (!_debug_cs_type){
 		return SLL_ACQUIRE_STATIC_INT(0);
 	}
 	sll_object_t* o=SLL_CREATE();
@@ -19,12 +20,12 @@ static sll_object_t* _debug_get_call_stack(const sll_object_t*const* al,sll_arg_
 		sll_object_t* nm=SLL_CREATE();
 		nm->t=SLL_OBJECT_TYPE_STRING;
 		sll_string_clone(sll_current_runtime_data->a_dt->st.dt+k->nm,&(nm->dt.s));
-		sll_object_t* dt[]={
+		sll_object_t* dt[3]={
 			nm,
 			SLL_FROM_INT(k->_ii),
 			SLL_FROM_INT(k->_s)
 		};
-		o->dt.a.v[i]=sll_create_object_type(sll_current_runtime_data->tt,_debug_type,dt,3);
+		o->dt.a.v[i]=sll_create_object_type(sll_current_runtime_data->tt,_debug_cs_type,dt,3);
 		SLL_RELEASE(nm);
 		SLL_RELEASE(dt[1]);
 		SLL_RELEASE(dt[2]);
@@ -79,11 +80,35 @@ static sll_object_t* _debug_get_ref_count(const sll_object_t*const* al,sll_arg_c
 
 
 
+static sll_object_t* _debug_get_vm_config(const sll_object_t*const* al,sll_arg_count_t all){
+	if (!_debug_vm_cfg_type){
+		return SLL_ACQUIRE_STATIC_INT(0);
+	}
+	sll_object_t* dt[4]={
+		SLL_FROM_INT(sll_current_vm_config->s_sz),
+		SLL_FROM_FILE(sll_current_vm_config->in),
+		SLL_FROM_FILE(sll_current_vm_config->out),
+		SLL_FROM_FILE(sll_current_vm_config->err)
+	};
+	sll_object_t* o=sll_create_object_type(sll_current_runtime_data->tt,_debug_vm_cfg_type,dt,4);
+	SLL_RELEASE(dt[0]);
+	SLL_RELEASE(dt[1]);
+	SLL_RELEASE(dt[2]);
+	SLL_RELEASE(dt[3]);
+	return o;
+}
+
+
+
 static sll_object_t* _debug_set_type(const sll_object_t*const* al,sll_arg_count_t all){
-	if (all){
+	if (all>1){
 		const sll_object_t* a=*al;
 		if (SLL_OBJECT_GET_TYPE(a)==SLL_OBJECT_TYPE_INT){
-			_debug_type=(sll_object_type_t)(a->dt.i);
+			_debug_cs_type=(sll_object_type_t)(a->dt.i);
+		}
+		a=*(al+1);
+		if (SLL_OBJECT_GET_TYPE(a)==SLL_OBJECT_TYPE_INT){
+			_debug_vm_cfg_type=(sll_object_type_t)(a->dt.i);
 		}
 	}
 	return SLL_ACQUIRE_STATIC_INT(0);
@@ -96,5 +121,6 @@ void _register_debug_functions(void){
 	sll_register_internal_function(sll_current_runtime_data->ift,SLL_CHAR("sll_ext:debug_get_instruction_count"),_debug_get_instruction_count,0);
 	sll_register_internal_function(sll_current_runtime_data->ift,SLL_CHAR("sll_ext:debug_get_name"),_debug_get_name,0);
 	sll_register_internal_function(sll_current_runtime_data->ift,SLL_CHAR("sll_ext:debug_get_ref_count"),_debug_get_ref_count,0);
+	sll_register_internal_function(sll_current_runtime_data->ift,SLL_CHAR("sll_ext:debug_get_vm_config"),_debug_get_vm_config,0);
 	sll_register_internal_function(sll_current_runtime_data->ift,SLL_CHAR("sll_ext:debug_set_type"),_debug_set_type,0);
 }